KTIJ-728 [Code Completion]: type arguments at wrong place for invoke()
Type arguments are calculated based on a synthetic PSI tree and later
applied to the target one when the user accepts one of the completion
variants. These trees are slightly different and the challenge is to
find the right place for insertion.

For calls of invoke() operator and regular function generated PSI trees
differ. Prior to this commit this fact was not taken into account and
completion followed invoke() operator failed.

^KTIJ-728 fixed

GitOrigin-RevId: e5fb25e4550cb332b022ec2ab1460e718b4d6042