KTIJ-18946 [New Project Wizard]: project driven module templates
The idea of this commit is to "inverse" the process of module templates
configuration.

So far a template itself decided whether it is included into suggested
options for a module or not, see [1]. The approach didn't match the new
requirement - ability to define applicable module templates
on a project level (including "none" case).

As a compromise solution, Template class is extended with
isPermittedForModule() function called prior to the logic described
above. The point is to check if Template matches expectations defined
for the Module on a project level. See [2].

---------------------------------------
[1]: Template.isSupportedByModuleType()
[2]: Module.permittedTemplateIds

GitOrigin-RevId: 6f4be77d1d1e2da1bd2bdd3d96f33de79fb86e15