KTIJ-19262 [New Project Wizard]: update modules/targets templates
All module templates of the wizard reside in a common registry.
To find module compliant templates algo traverses the latter
asking each template whether it's suitable for the module
in question. Templates decide but not fully.

Starting from 6f4be77d the concept of "project template decides"
was introduced. Modules/targets declared as a part of
project-template can be supplied with a list of permitted-templates.

This commit updates permitted module templates for the set
of project ones.

GitOrigin-RevId: 36fd70664c105dc3e5050bef78ac48926d68d271