[kotlin] "Convert argument to Set" inspection/intention
Convert argument to Set to improve performance: an intention and an
inspection in preparation to Kotlin 1.6.
    
The inspection detects the calls to several functions on arrays and
iterables that call `contains` on their argument (e.g., minus or
intersect), and proposes a quick fix to convert their argument
into a set if it may improve performance.
    
^KTIJ-19104 Fixed
^KTIJ-19106 Fixed

GitOrigin-RevId: 6042ddaa9513e893a4aa3b250cf19d4e43f0cc1c