Limit safe-push params for importing tests
Safe Push isn’t supposed to check everything, it is supposed
to prevent pushing completely broken commits into master.
So we decided to do not run all versions of Gradle and
Kotlin Gradle Plugin for importing tests. For safe-push tests
leaving only the master version of KGP and 6.8.2 of Gradle

GitOrigin-RevId: 1905e673c4c0745fc7310067d57ba0993026dd90