[kotlin] never step to fake line numbers generated by the Kotlin compiler.
The compiler inserts a fake line number for single-line inline
function calls with a lambda parameter such as:

   `42.let { it + 1 }`

This is done so that a breakpoint can be set on the lambda and
on the line even though the lambda is on the same line. When
stepping, we do not want to stop at such fake line numbers.
They cause us to step to line 1 of the current file.

For StepOver this is simple as we just need to filter them out
in the existing Kotlin specific debugger code.

For StepInto we need to add special Kotlin handling.

^KTIJ-18645 Fixed.

(cherry picked from commit 79e85397ad4dd185dadd17aa8999d91ae8c4b1f9)

IJ-MR-15469

GitOrigin-RevId: 0cd994ae883edf792d29a724b6de7a7ec1486318