[kotlin] Compute DeferredTypes only created inside the BodyResolver (KTIJ-4857)
BodyResolver triggers computation of all pending DeferredTypes that settle in the current BindingTrace. As all major implementations of BindingTrace return both own keys and keys of its owner, all pending types in the whole trace are computed.

However, BindingTrace can be used by several threads in parallel, as we reuse the single container for all 'analyze()' calls (see PerFileAnalysisCache). As a result, we can easily get a concurrent modification inside a loop.

The fix forces the BodyResolver to handle only its own keys for the DEFERRED_TYPE slice (and it's the only place where DEFERRED_TYPE is currently used).

GitOrigin-RevId: d445cf82a9b4b44ae8b8820f57699e16506b2e39