[kotlin] wrap evaluated code fragment analysis in non-blocking read action
Sometimes analysis of an evaluated code fragment can be interrupted with the ProcessCancelledException. In this case a user will see the exception instead of an evaluation result. So we should wrap the analysis in the non-blocking read action, so it will be rerun if ProcessCancelledException is thrown.

^KTIJ-19974 Fixed

(cherry picked from commit c63170f36b2b49652e30e347409e35081f2e2f8d)

IJ-MR-15796

GitOrigin-RevId: a672e49b6edbb403418c08ef847f4965acbfdeff