[kotlin j2k] Fix string concatenation with multiple leading non-String operands
The original bug in KTIJ-16105 description was already obsolete.

This is a fix of a more specific case from
https://youtrack.jetbrains.com/issue/KTIJ-16105#focus=Comments-27-5294439.0-0

Before the fix, the first subexpression `1 + 1` inside an expression like `1 + 1 + "1"`
had the wrong type `String`, which is the type of the entire `PsiPolyadicExpression`.
Now, we compute its proper type `Int` and append a `toString()` call accordingly.

The case with multiple Chars `'c' + 'd' + "5"` still doesn't work due to
https://youtrack.jetbrains.com/issue/KTIJ-8532

^KTIJ-16105 Fixed

GitOrigin-RevId: c09df62414d4aa83bff8442475765b252f67691f