KTIJ-9370 Deprioritize `suspend` keyword in completion
In places like `fun foo(a: <caret>)` or `fun foo(a: List<<caret>>)`
we don't often need `suspend` keyword; certainly not often enough
for it to be higher than every other possible type
(since `keyword` kind > `default` kind).

This commit fixes that problem; it adds a kind called
`notProbableKeyword` (in contrast with `probableKeyword`) and applies
it to the `suspend` keyword lookup item when it is completed inside
`KtTypeReference` element.

Note that if somebody still wants the `suspend` keyword in the type position,
(e.g. in function like `fun foo(action: suspend () -> Unit)`),
typing single 's' letter as a prefix will filter 99% of
types (because the completion in the type position is case-sensitive)

^KTIJ-9370 Fixed

GitOrigin-RevId: 2f81616e6ebf1beaf1ee30842f2700afab0f8f6c