[kotlin] Remove compiler-components-for-jps jar dependency
Several modules have dependency on kotlinc.compiler-components-for-jps
library which resolves into binary artifact
compiler-components-for-jps-1.5.10-release-955.jar

This artifact was added during migration from `kotlin` to `intellij`
repository and contains classes from `kotlin.jps-common`module.
For this reason classpath of Kotlin IDE plugin contains duplicated
class files and accidentally chooses ones from jar.

Deleting of this dependency seems to be safe because this dependency
was not ever using anywhere in plugin.

#KTIJ-19864 Fixed
#KTIJ-20076 Fixed
#KTIJ-20086 Fixed

GitOrigin-RevId: cde81e8978a45a9d486dc85bc2770933b5dd77ea