[Kotlin] Do shallow search of stdlib in module dependencies
Module info dependencies are not transitive, and therefore
search for Kotlin standard library dependency (for builtins)
used to be done in a transitive closure as a safety measure.
Such a search is costly, especially for large projects with
lots of modules and libraries. However, there is no evident
cases when transitive stdlib dependency would be the expected
one for correctly configured project. Apparently, it might
happen only in two cases:
- required dependency is not added to a source module, while
some of its direct dependencies has it (misconfiguration)
- library dependency on other library is a result of approximation
from the IJ workspace model and the original library doesn't
have it; then not adding such a dependency on a client module
is totally reasonable.

Furthermore, specifically Kotlin stdlib is eagerly added to
most of the modules; searching for transitive stdlib instead
of skipping and falling back to DefaultBuiltins for such
modules also seems incorrect.

Relevant tests:
MultiplatformAnalysisTestGenerated#testBuiltinsAndStdlib
MultiplatformAnalysisTestGenerated#testUseCorrectBuiltIns

Relates to:
KTIJ-11618
KT-33233

GitOrigin-RevId: 3bc9a08410564ca29e9cf73c748c7af9c040326b