[kotlin] Basic highlighting of definitely non-nullable types
This commit makes definitely non-nullable types `T & Any` (where `T` is
a type parameter) to be hightlighted as a whole. Previously only `T` was
hightlighted as a type parameter.

The type `T & Any` is a non-nullable subtype of the type parameter `T`,
so the type parameter hightlighting is used for the entire type.

^KTIJ-20253 Fixed

GitOrigin-RevId: 619cdb60a8356cc0513c5c71049f5666fc95111a