[Kotlin] Fix AndroidQuickFixTest
With HMPP relation between source sets and targets is changed
in the way that common source sets might have less targets
that before HMPP. In case of single-target project
commonMain source set will have only one target.

AndroidQuickFixTest are testing behaviour of creating actuals
for single-android-target project. Prior HMPP it was expected
that actuals would be created in android-only source sets.
With HMPP this behaviour is changed a bit, due to the fact that
commonMain is already android-only source set. Therefore actuals
got created in commonMain.

This might not be a desrired behaviour. Decision can be found
in related issue: KTIJ-20424

As part of the fix of AndroidQuickFixTest separate `afterHmpp`
directory is introduced to place expected code
after "quick fix" for project with enabled HMPP.

GitOrigin-RevId: fb97d8cdef3e0d01103d5725fa0432549a5540dd