[kotlin] Resolve common code from JVM perspective for evaluator
Under composite mode, common code is resolved to common
descriptors/dependencies. If JVM evaluator is stopped at the breakpoint
in common code, it will resolve everything (e.g. evaluated expression)
in context of common, and will get common resolution results. Then it
will blow up, because JVM evaluator expects the resolution results to be
JVM

The idea of the fix is to essentially provide a non-composite facade
(i.e. the way how common code used to be resolved for JVM before HMPP)
specifically for evaluator

^KTIJ-20716 Fixed

GitOrigin-RevId: e6cbf8c3ee94d22c8aba6668e75c22d8fae89af4