[kt-*-only] Fixed test KotlinUastTypesTest#testAnnotatedTypes
There is 2 changes. In testdata -- it is because of f08b2c8e28
And the second one. The story behind that is strange.
Some time ago there were 2 copy of TypesTestBase, and in
e430a1ea14d5a90 it was fixed. But in parallel branch there was
commit e8f294316c5d97cf6 and that change was lost.

But because now the code of TypesTestBase is in IDEA platform, we
cannot change it in kt-* branch.

Please do deduplication again in IDEA master :)

GitOrigin-RevId: f1fa42b04329a0beb646565c563fc8cc354e3cd5