Additional changes from [import] Fix Kotlin facet creation with convention plugin
In kt-212-master we have the same commit, but a bit different version
of it: 277c81a43bd91f16a and this version was added into kt-213-*
branches.

In 213 -- we have "the full version" under aedcfdddef9b544fddee6f6b.

So this commit just adds the missing changes from 213 branch

GitOrigin-RevId: dbef1e81718780b9be3a57a12e7245b4adcea620