KTIJ-20787 MPP: run icon missed for common tests
This commit fixes the case when MPP project
- has only two targets: JVM & Android.
- has HMPP mode on (aka kotlin.mpp.enableGranularSourceSetsMetadata).
- kotlin.test.Test annotation is used to mark tests.

[What was wrong]
* Depending on HMPP mode common-module platform is calculated differently [1].
When it's off the platform is always COMMON eventually having three targets:
(jvm, js, native). Being off means that common-source-set compilations are
analysed honestly: (JVM + Android) results in JVM-platform.
* Test run icons are platform dependent. Should a common-module have COMMON
platform its test classes icons are built with [2], in case of JVM it's
[3]. [2] is a composite containing [3] which didn't actually work in this
case as well - JS/Native is what made the trick.
* JVM-tooling can resolve @Test annotation based on typealiases only. It should
be an exact class residing in module classpath (e.g. org.junit.Test for JUnit4).
* When a project has JVM or Android target (but not both), it works fine. Having
them together results in a collision: Android depends on JUnit4 whereas JVM has
JUnit5 as a default [4]. Hence, no dependency on jar with typealias, which is ok.

[How it's fixed]
The fix includes two points: JVM-tooling got a fallback to generate so-called
generic icons (used for JS-modules) when @Test cannot be resolved. Configuration
producers for common modules are now applicable not only to COMMON platform but
to JVM one as well.

[Tests considered]
- MultiModuleLineMarkerTestGenerated
- LineMarkersTestGenerated

------------------------------------------------------------------------------
[1]: KotlinMPPGradleModelBuilder.computeSourceSetPlatforms
[2]: CommonIdePlatformKindTooling.getTestIcon
[3]: JvmIdePlatformKindTooling.getTestIcon
[4]: KotlinDependenciesManagement.kotlinTestCapabilityForJvmSourceSet (kotlin)

GitOrigin-RevId: c2b79898df671acf167e39f12d864f752c6337fa