[kotlin j2k] Fix optimize imports post-processing
The import optimizer is run in two steps
(see AbstractLayoutCodeProcessor#performFileProcessing):

Step 1: In read action: collect used imports
Step 2: In write action: remove unused imports from the file

The problem is that step 2 is executed asynchronously on the EDT,
when the original KtFile is no longer relevant: its contents
have been copied to a new KtFile in JavaToKotlinAction#convertFiles.
So, it correctly optimizes imports, but in the wrong file.

To fix the bug, we bypass the general IDEA infrastructure for
optimizing imports through `OptimizeImportsProcessor` and instead
directly invoke `KotlinImportOptimizer` in place, synchronously.

Note: the problem doesn't reproduce in test mode, because there,
both steps are executed on the EDT synchronously, one after another.

^KTIJ-14272 Fixed

GitOrigin-RevId: a545e05586247e10016d7e8b441043da81025086