[kotlin] Update test data for 'IntentionTestGenerated', expression evaluator
Since Kotlin 1.7, frontend doesn't analyze complex boolean expressions anymore (KT-39883). 'SimplifyBooleanWithConstantsInspection' depends on it, so several cases broke.

The updated test data signifies the tests should run with LV 1.6. Obviously, it's hardly a solution, but rather a way to mute the tests. So here's KTIJ-21274 to track the actual problem.

GitOrigin-RevId: 135a785796759fb4349b0e21b9dc0956f4d45354