Update bundled Kotlin compiler version (1.6.20-RC, KTIJ-21283)
This commits re-bundled the same artifacts with the corrected version. In 'b662c390d829d4a9d741de955b0b3ce9a53055ca', compiler components were updated to '1.6.20-235' (which is stable version). Various places broke because of that, as it's technically an RC, and stable artifacts are absent in Maven Central (Kotlin 1.6.20 is not released yet).

GitOrigin-RevId: 35afdcb500b9b25d3e6c8cab871cfbed26d1aebd