[kotlin] Support intermediate checks in 'VerboseNullabilityAndEmptinessInspection' (KTIJ-20207, KTIJ-13836)
As there are no polyadic expressions in Kotlin, multiple logical checks become nested in PSI. Previous implementation of 'VerboseNullabilityAndEmptinessInspection' assumed both nullability checks are always in left-right positions of the same binary expression. However, it's not always the case.

For instance, the expression 'flag && a != null && a.isNotEmpty()' is parsed as '[flag && a != null] && a.isNotEmpty()'. This commit modifies the inspection, so it recognizes such cases.

GitOrigin-RevId: 5961adaf25e11b6a581c1d551bffeb0eb2565c48