[kotlin] Update Maven/Gradle configurator tests, add actual tests
Before, the bundled compiler version were used in most tests. As the bundled version is different in 'master' and 'kt-master', adjustments had to be made each master/kt-master merge.

Now the version is fixed in all tests, so all tests verify the same things in both branches. In additions, more test variations are added for different version patterns, such as '-SNAPSHOT' or '-dev'.

GitOrigin-RevId: 5f538c75cbe97625d09a0c48dcf1aaba7cff5179